rename crawler workers
