+++ app/jobs/course_sync_job.rb
@@ -1,0 +1,14 @@
+require 'thread'
+require 'thwait'
+require 'set'
+
+class CourseSyncJob
+  include Sidekiq::Worker
+  sidekiq_options retry: false
+
+  def perform(*args)
+    Course.import_to_course(*args)
+
+    CourseCacheGenerateJob.perform_later(*args)
+  end
+end
