+++ lib/tasks/course_crawler.rake
@@ -2,8 +2,8 @@ namespace :course_crawler do
 
   desc 'Run up course crawler directly, ex: rake course_crawler:run[ntust,2015,2,true,false]'
   task :run, [:organization_code, :year, :term, :save_json, :save_to_db] => :environment do |_, args|
-    CourseCrawler::CourseWorker.new.perform(
-      "#{args.organization_code.camelize}CourseCrawler",
+    CourseCrawlerJob.new.perform(
+      args.organization_code,
       year:       args.year && !args.year.to_i.zero? && args.year.to_i,
       term:       args.term && !args.term.to_i.zero? && args.term.to_i,
       save_json:  true || args.save_json == 'true',
